Revert "post/Makefile: Only build FP post tests if enabled via CONFIG_SYS_POST_FPU"
This reverts commit 46b75403dd301d476f51bfab1184e65d8220f3da.

The logic of this patch is broken - testing for CONFIG_SYS_POST_FPU in
the Makefile cannot work, as this is only a bit that may (or may not)
be set in the CONFIG_POST variable.

The patch cases build errors on a number of boards, so we revert it.

Signed-off-by: Wolfgang Denk <wd@denx.de>