i.MX28: Fix initial stack pointer position
The patch:

  m28evk: Use GENERATED_GBL_DATA_SIZE
  commit 1a1c7b5d61cf7671cd860e993cd323392b352a75

introduced usage of GENERATED_GBL_DATA_SIZE and calculation of initial stack
pointer position defived from that. Due to a small typo, the SP position moved
to 0x21f80, which is past the SRAM area. This didn't manifest on the real
hardware as the SRAM repeats there (address bits in the CPU being ignored).

Though this was caught in QEMU, where it crashed the emulator.

Signed-off-by: Marek Vasut <marex@denx.de>
Cc: Stefano Babic <sbabic@denx.de>
Cc: Fabio Estevam <fabio.estevam@freescale.com>
Tested-by: Fabio Estevam <fabio.estevam@freescale.com>